% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjectives}
\alias{adjectives}
\title{A Data Frame of Adjectives}
\format{
a data frame with 1,347 observations and 1 column.
\describe{
\item{\code{value}}{a character vector}
}
}
\usage{
adjectives
}
\description{
This is a data frame of adjectives in the English language to use as an attribute in generating a code name.
}
\details{
Adjectives found on Github as a gist from username "@hugsy".
}
\keyword{datasets}
