% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animals}
\alias{animals}
\title{A Data Frame of Animals}
\format{
a data frame with 400 observations and 1 column.
\describe{
\item{\code{value}}{a character vector}
}
}
\usage{
animals
}
\description{
This is a data frame of animals in the English language to use as an object in generating a code name.
}
\details{
Animals found on Github as a gist from username "@atduskgreg".
}
\keyword{datasets}
