% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codename_message.R
\name{codename_message}
\alias{codename_message}
\title{Display package version for \pkg{codename}}
\usage{
codename_message()
}
\value{
\code{codename_message()} produces a message about the installed version of \pkg{codename}. Successive updates
may (understandably) break an expected output from a reproducible seed in the \code{codename()} function. This just
adds an extra layer of transparency.
}
\description{
\code{codename_message()} produces a message about the package version.
}
\examples{

codename_message()

}
\author{
Steven V. Miller
}
