% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gods}
\alias{gods}
\title{A Data Frame of Gods}
\format{
a data frame with 221 observations and 1 column.
\describe{
\item{\code{value}}{a character vector}
}
}
\usage{
gods
}
\description{
This is a data frame of gods in the English language to use as an object in generating a codename.
}
\details{
Gods data cobbled from the website "Godchecker" after searching for the top 10 most popular deities by various regions/religions on their website.
}
\keyword{datasets}
