% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nouns}
\alias{nouns}
\title{A Data Frame of Nouns}
\format{
a data frame with 6801 observations and 1 column.
\describe{
\item{\code{value}}{a character vector}
}
}
\usage{
nouns
}
\description{
This is a data frame of nouns in the English language to use as an object in generating a codename.
}
\details{
Nouns data come by way of "The Great Nouns List"
}
\keyword{datasets}
