% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{st_chars}
\alias{st_chars}
\title{A Data Frame of Star Trek character names}
\format{
a data frame with 5745 observations and 1 column.
\describe{
\item{\code{value}}{a character vector}
}
}
\usage{
st_chars
}
\description{
This is a data frame of characters from the Star-trek universe to use as an object in generating a code name.
}
\details{
data retrieved using stapi from the rtrek pkg.
}
\keyword{datasets}
