% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variety_pack.R
\name{variety_pack}
\alias{variety_pack}
\title{Get a variety pack of unique code names for yourself, your organization, other people, your projects, and whatever else}
\usage{
variety_pack(seed)
}
\arguments{
\item{seed}{an optional reproducible seed, which can be specified as a character or number.}
}
\value{
\code{variety_pack()} takes an optional reproducible seed and produces all kinds of code names available in
\pkg{codename}.
}
\description{
\code{variety_pack()} is produces all types of code names available in \pkg{codename}.
}
\examples{

variety_pack()
variety_pack(seed = 8675309)

}
\author{
Steven V. Miller
}
