% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wu_adjs}
\alias{wu_adjs}
\title{A Data Frame of Adjectives from the "Wu-Tang Name Generator"}
\format{
a data frame with 45 observations and 1 column.
\describe{
\item{\code{value}}{a character vector}
}
}
\usage{
wu_adjs
}
\description{
This is a data frame of adjectives from the "Wu-Tang Name Generator" to use as an attribute in generating a code name.
}
\details{
These data are classic and apparently come from around 2002.
}
\keyword{datasets}
