% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wu_nouns}
\alias{wu_nouns}
\title{A Data Frame of Nouns from the "Wu-Tang Name Generator"}
\format{
a data frame with 40 observations and 1 column.
\describe{
\item{\code{value}}{a character vector}
}
}
\usage{
wu_nouns
}
\description{
This is a data frame of nouns from the "Wu-Tang Name Generator" to use as an object in generating a codename.
}
\details{
These data are classic and apparently come from around 2002.
}
\keyword{datasets}
