% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xkcd_colors}
\alias{xkcd_colors}
\title{A Data Frame of Colors}
\format{
a data frame with 949 observations and 1 column.
\describe{
\item{\code{value}}{a character vector}
}
}
\usage{
xkcd_colors
}
\description{
This is a data frame of colors in the English language to use as an attribute in generating a code name.
}
\details{
Colors found by way of the web comic \emph{XKCD}.
}
\keyword{datasets}
