% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codename.R
\name{codename}
\alias{codename}
\title{Generate a unique codename for yourself, your organization, other people, your projects, and whatever else}
\usage{
codename(type = "any", seed)
}
\arguments{
\item{type}{a type of code the user wants. Defaults to "any", but "gods", "ubuntu", and "wu-tang" are available.}

\item{seed}{an optional reproducible seed, which can be specified as a character or number.}
}
\value{
\code{codename()} takes a preferred type of code and an optional reproducible seed and returns
a codename for the user to consider for whatever it is they want.
}
\description{
\code{codename()} is a tool for generating codenames for various things.
}
\examples{

codename()
codename(type = "ubuntu")
codename(type = "gods")
codename(type = "wu-tang")
codename(seed = 8675309)
codename(seed = 8675309)
codename(seed = "a character")
codename(seed = "a character")

}
\author{
Steven V. Miller
}
